changed navbar link. added institution end-point. can change profile pic and edit model
