+++ views/institution/finance.ejs
@@ -1,0 +1,76 @@
+<% include ../partials/header %>
+<% include ../partials/institution-sidebar %>
+<% include ../partials/institution-navbar %>
+
+<div class="content">
+    <div class="container-fluid">
+        <div class="row">
+            <div class="col-md-12">
+                <div class="card">
+                    <div class="card-header" data-background-color="purple">
+                        <h4 class="title">Finance Committee</h4>
+                        <p class="category">information about duration, designation etc.</p>
+                    </div>
+                    <div class="card-content table-responsive">
+                        <table class="table">
+                            <thead class="text-primary">
+                                <th>Sl. No</th>
+                                <th>Name</th>
+                                <th>Category</th>
+                                <th>Address</th>
+                                <th>Status</th>
+                            </thead>
+                            <tbody>
+                                <tr>
+                                    <td>Dakota Rice</td>
+                                    <td>Niger</td>
+                                    <td>Oud-Turnhout</td>
+                                    <td class="text-primary">$36,738</td>
+                                    <td>Niger</td>
+                                </tr>
+                                <tr>
+                                    <td>Minerva Hooper</td>
+                                    <td>Curaçao</td>
+                                    <td>Sinaai-Waas</td>
+                                    <td class="text-primary">$23,789</td>
+                                    <td>Niger</td>
+                                </tr>
+                                <tr>
+                                    <td>Sage Rodriguez</td>
+                                    <td>Netherlands</td>
+                                    <td>Baileux</td>
+                                    <td class="text-primary">$56,142</td>
+                                    <td>Niger</td>
+                                </tr>
+                                <tr>
+                                    <td>Philip Chaney</td>
+                                    <td>Korea, South</td>
+                                    <td>Overland Park</td>
+                                    <td class="text-primary">$38,735</td>
+                                    <td>Niger</td>
+                                </tr>
+                                <tr>
+                                    <td>Doris Greene</td>
+                                    <td>Malawi</td>
+                                    <td>Feldkirchen in Kärnten</td>
+                                    <td class="text-primary">$63,542</td>
+                                    <td>Niger</td>
+                                </tr>
+                                <tr>
+                                    <td>Mason Porter</td>
+                                    <td>Chile</td>
+                                    <td>Gloucester</td>
+                                    <td class="text-primary">$78,615</td>
+                                    <td>Niger</td>
+                                </tr>
+                            </tbody>
+                        </table>
+                    </div>
+                    <button type="submit" class="btn btn-primary pull-right" style="margin:13px">Edit</button>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+
+<% include ../partials/footer %>
