+++ views/institution/governance.ejs
@@ -1,6 +1,6 @@
 <% include ../partials/header %>
-<% include ../partials/faculty-sidebar %>
-<% include ../partials/faculty-navbar %>
+<% include ../partials/institution-sidebar %>
+<% include ../partials/institution-navbar %>
 
 <div class="content">
     <div class="container-fluid">
@@ -8,7 +8,7 @@
             <div class="col-md-12">
                 <div class="card">
                     <div class="card-header" data-background-color="purple">
-                        <h4 class="title">Academic Committee</h4>
+                        <h4 class="title">Governance Committee</h4>
                         <p class="category">information about duration, designation etc.</p>
                     </div>
                     <div class="card-content table-responsive">
