+++ views/partials/faculty-navbar.ejs
@@ -13,21 +13,10 @@
           <div class="collapse navbar-collapse">
               <ul class="nav navbar-nav navbar-right">
                 <li class="dropdown">
-                    <a href="#" class="dropdown-toggle" data-toggle="dropdown">
+                    <a href="/institution">
                       <i class="material-icons hidden-lg hidden-md">people</i>
-                      <p>Commitiees</p>
+                      <p>Institution</p>
                     </a>
-                    <ul class="dropdown-menu">
-                        <li>
-                            <a href="/commitee/governance">Governance</a>
-                        </li>
-                        <li>
-                            <a href="/commitee/finance">Finance</a>
-                        </li>
-                        <li>
-                            <a href="/commitee/academic">Academic</a>
-                        </li>
-                    </ul>
                 </li>
                   <li class="dropdown">
                       <a href="#" class="dropdown-toggle" data-toggle="dropdown">
