+++ views/partials/hod-navbar.ejs
@@ -1,6 +1,6 @@
 </div>
 <div class="main-panel">
-  <nav class="navbar navbar-transparent navbar-fixed-top">
+  <nav class="navbar navbar-fixed-top" style="background-color:#EEEEEE; z-index:3">
       <div class="container-fluid">
           <div class="navbar-header">
               <button type="button" class="navbar-toggle" data-toggle="collapse">
@@ -12,6 +12,35 @@
           </div>
           <div class="collapse navbar-collapse">
               <ul class="nav navbar-nav navbar-right">
+                <li class="dropdown">
+                    <a href="/institution">
+                      <i class="material-icons hidden-lg hidden-md">people</i>
+                      <p>Institution</p>
+                    </a>
+                </li>
+                  <li class="dropdown">
+                      <a href="#" class="dropdown-toggle" data-toggle="dropdown">
+                        <i class="material-icons hidden-lg hidden-md">store</i>
+                        <p>Department</p>
+                      </a>
+                      <ul class="dropdown-menu">
+                          <li>
+                              <a href="/department/cse">CSE</a>
+                          </li>
+                          <li>
+                              <a href="/department/ise">ISE</a>
+                          </li>
+                          <li>
+                              <a href="/department/iem">IEM</a>
+                          </li>
+                          <li>
+                              <a href="/department/mech">MECH</a>
+                          </li>
+                          <li>
+                              <a href="/department/ece">ECE</a>
+                          </li>
+                      </ul>
+                  </li>
                   <li class="dropdown">
                       <a href="#" class="dropdown-toggle" data-toggle="dropdown">
                           <i class="material-icons">notifications</i>
@@ -41,6 +70,14 @@
                           <i class="material-icons">person</i>
                           <p class="hidden-lg hidden-md">Profile</p>
                       </a>
+                      <ul class="dropdown-menu">
+                        <li>
+                            <a href="/faculty">Profile</a>
+                        </li>
+                        <li>
+                            <a href="/v1/apis/logout">Logout</a>
+                        </li>
+                      </ul>
                   </li>
               </ul>
               <form class="navbar-form navbar-right" role="search">
