--- views/partials/users-sidebar.ejs
@@ -1,58 +1,0 @@
-<div class="sidebar" data-color="purple" data-image="../assets/img/sidebar-1.jpg" style="z-index:4">
-
-<div class="logo" style=" background-color:#9C27B0">
-    <img src="/assets/img/brand.png" style="height:70px;" class="img-responsive">
-    </img>
-</div>
-<div class="sidebar-wrapper">
-    <ul class="nav">
-        <li class="active">
-            <a href="/department">
-                <i class="material-icons">dashboard</i>
-                <p>General Information</p>
-            </a>
-        </li>
-        <li>
-            <a href="/department/qualification">
-                <i class="material-icons">person</i>
-                <p>Student Information</p>
-            </a>
-        </li>
-        <li>
-            <a href="/department/service-details">
-                <i class="material-icons">content_paste</i>
-                <p>Infrastructure Details</p>
-            </a>
-        </li>
-        <li>
-            <a href="/department/achievements">
-                <i class="material-icons">library_books</i>
-                <p>Admission Details</p>
-            </a>
-        </li>
-        <li>
-            <a href="/department/academic-details">
-                <i class="material-icons">bubble_chart</i>
-                <p>Activities</p>
-            </a>
-        </li>
-        <li>
-            <a href="/department/rnd-details">
-                <i class="material-icons">location_on</i>
-                <p>BOS | BOE</p>
-            </a>
-        </li>
-        <li id="personnallist">
-            <a href="/department/personnal-info">
-                <i class="material-icons text-gray">dashboard</i>
-                <p>Faculty Information</p>
-            </a>
-        </li>
-        <li class="active-pro">
-            <a href="upgrade.html">
-                <i class="material-icons">unarchive</i>
-                <p>Upgrade to PRO</p>
-            </a>
-        </li>
-    </ul>
-</div>
